server/requirements.txt
unstructured
tiktoken
pinecone-client
pinecone[grpc]
langchain-google-genai
langchain
python-dotenv
langchain-community
flask
flask-cors